% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSoilPars.R
\name{getPointCoordinates}
\alias{getPointCoordinates}
\title{Title getPointCoordinates}
\usage{
getPointCoordinates(geoLaenge, geoBreite, crs = "EPSG:25832")
}
\arguments{
\item{geoLaenge}{the longitude of the point}

\item{geoBreite}{the latitude of the point}

\item{crs}{the coordinate reference system of the point}
}
\value{
an sf object with the point coordinates in the given crs
}
\description{
Title getPointCoordinates
}
