% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSoilPars.R
\name{getSoilMap}
\alias{getSoilMap}
\title{Title getSoilMap}
\usage{
getSoilMap(point_coordinates, BUEK2000_Kartenblaetter, BUEK2000_path)
}
\arguments{
\item{point_coordinates}{the coordinates of the point as an sf object}

\item{BUEK2000_Kartenblaetter}{the map sheets of the BUEK2000 data}

\item{BUEK2000_path}{the path to the BUEK2000 data}
}
\value{
the selected map sheet of the BUEK2000 data as sf multipolygon object
}
\description{
Title getSoilMap
}
