% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSoilPars.R
\name{getSoilPolygon}
\alias{getSoilPolygon}
\title{Title getSoilPolygon}
\usage{
getSoilPolygon(point_coordinates, Kartenblatt_spatial, BUEK2000_code)
}
\arguments{
\item{point_coordinates}{the point coordinates as an sf object}

\item{Kartenblatt_spatial}{the selected map sheet of the BUEK2000 data}

\item{BUEK2000_code}{the BUEK2000 code data}
}
\value{
a data frame with the soil polygon data including texture data for the soil profiles
of the "Leitboden", i.e. the most important soil profile and additional "Begleitböden"
}
\description{
Title getSoilPolygon
}
