% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSoilPars.R
\name{getSoilTexture}
\alias{getSoilTexture}
\title{Title getSoilTexture}
\usage{
getSoilTexture(soil_polygon, short = TRUE, fill_NA = TRUE)
}
\arguments{
\item{soil_polygon}{the soil polygon data including texture data for the soil profiles}

\item{short}{if TRUE, the data frame is reduced to the basic information}

\item{fill_NA}{if TRUE, the texture is set to the value of the upper horizon if texture is NA}
}
\value{
a data frame with the soil texture data for the soil profile of the "Leitboden",
}
\description{
Title getSoilTexture
}
