% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSoilPars.R
\name{getSoilMap}
\alias{getSoilMap}
\title{Title getSoilMap}
\usage{
getSoilMap(point_coordinates)
}
\arguments{
\item{point_coordinates}{the coordinates of the point as an sf object}
}
\value{
the selected map sheet of the BUEK2000 data as sf multipolygon object
}
\description{
Title getSoilMap
}
